requirements.txt
absl-py==1.0.0
amqp==5.1.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
billiard==3.6.4.0
cachetools==5.0.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.73.0
flatbuffers==2.0
flower==1.0.0
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.6.0
httptools==0.3.0
humanize==4.0.0
idna==3.3
importlib-metadata==4.11.0
JPype1==1.3.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kombu==5.2.4
libclang==13.0.0
Markdown==3.3.6
mine-replayed @ file:///Users/juan/workspace/python/engine/rid-fastapi/whl/mine_replayed-0.0.4-py3-none-any.whl
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
Pillow==9.0.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
supervisor==4.2.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tornado==6.1
typing_extensions==4.1.0
urllib3==1.26.8
uvicorn==0.17.4
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
